refactor: move evm factory in mod evm
